move tests to e2e folder
