+++ internal/controller/dbroleclaim_controller_test.go
@@ -18,7 +18,6 @@ package controller
 
 import (
 	"context"
-	"strconv"
 	"testing"
 
 	"github.com/aws/smithy-go/ptr"
@@ -37,8 +36,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
-	basefun "github.com/infobloxopen/db-controller/pkg/basefunctions"
-	. "github.com/infobloxopen/db-controller/pkg/dbclient"
 	"github.com/infobloxopen/db-controller/pkg/roleclaim"
 	. "github.com/infobloxopen/db-controller/testutils"
 	_ "github.com/lib/pq"
@@ -116,12 +113,12 @@ func TestReconcileDbRoleClaim_CopyExistingSecret(t *testing.T) {
 			},
 		}
 
-		controllerReconciler.reconciler = &roleclaim.DbRoleClaimReconciler{
+		controllerReconciler.Reconciler = &roleclaim.DbRoleClaimReconciler{
 			Client: controllerReconciler.Client,
 			Config: controllerReconciler.Config,
 		}
 
-		_, err := controllerReconciler.reconciler.Reconcile(ctx, reconcile.Request{
+		_, err := controllerReconciler.Reconciler.Reconcile(ctx, reconcile.Request{
 			NamespacedName: typeNamespacedName,
 		})
 		Expect(err).NotTo(HaveOccurred())
@@ -194,269 +191,16 @@ func TestSchemaUserClaimReconcile_WithNewUserSchemasRoles_MissingParameter(t *te
 				Config: tt.rec.Config,
 			}
 
-			r.reconciler = &roleclaim.DbRoleClaimReconciler{
+			r.Reconciler = &roleclaim.DbRoleClaimReconciler{
 				Client: r.Client,
 				Config: r.Config,
 			}
 
-			_, err := r.reconciler.Reconcile(tt.rec.Context, tt.rec.Request)
+			_, err := r.Reconciler.Reconcile(tt.rec.Context, tt.rec.Request)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("error = %v, wantErr %v", err, tt.wantErr)
 				return
 			}
-		})
-	}
-}
-
-func TestSchemaUserClaimReconcile_WithNewUserSchemasRoles(t *testing.T) {
-	RegisterFailHandler(Fail)
-	logf.SetLogger(zap.New(zap.WriteTo(GinkgoWriter), zap.UseDevMode(true)))
-
-	testDB := SetupSqlDB(t, "user_a", "masterpassword")
-	defer testDB.Close()
-
-	type reconciler struct {
-		Client             client.Client
-		Log                logr.Logger
-		Scheme             *runtime.Scheme
-		Config             *roleclaim.RoleConfig
-		DbIdentifierPrefix string
-		Context            context.Context
-		Request            controllerruntime.Request
-	}
-
-	viperObj := viper.New()
-	viperObj.Set("passwordconfig::passwordRotationPeriod", 60)
-
-	tests := []struct {
-		name    string
-		rec     reconciler
-		wantErr bool
-	}{
-		{
-			"Get UserSchema claim 1",
-			reconciler{
-				Client: &MockClient{Port: strconv.Itoa(testDB.Port)},
-				Config: &roleclaim.RoleConfig{
-					Viper: viperObj,
-					Class: "default",
-				},
-				Request: controllerruntime.Request{
-					NamespacedName: types.NamespacedName{Namespace: "schema-user-test", Name: "schema-user-claim-1"},
-				},
-				Log: zap.New(zap.UseDevMode(true)),
-			},
-			false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			r := &DbRoleClaimReconciler{
-				Client: tt.rec.Client,
-				Config: tt.rec.Config,
-			}
-
-			r.reconciler = &roleclaim.DbRoleClaimReconciler{
-				Client: r.Client,
-				Config: r.Config,
-			}
-
-			result, err := r.reconciler.Reconcile(tt.rec.Context, tt.rec.Request)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			Expect(result.Requeue).Should(BeFalse())
-
-			existingDBConnInfo, err := persistancev1.ParseUri(testDB.URL())
-			if err != nil {
-				t.Errorf("error = %v", err)
-				return
-			}
-
-			dbClient, err := basefun.GetClientForExistingDB(existingDBConnInfo, &controllerruntime.Log)
-			if err != nil {
-				t.Errorf("error = %v", err)
-				return
-			}
-
-			var responseUpdate = r.Client.(*MockClient).GetResponseUpdate()
-			Expect(responseUpdate).Should(Not(BeNil()))
-			var schemaUserClaimStatus = responseUpdate.(*persistancev1.DbRoleClaim).Status
-			Expect(schemaUserClaimStatus).Should(Not(BeNil()))
-			Expect(schemaUserClaimStatus.Error).Should(BeEmpty())
-			Expect(schemaUserClaimStatus.SchemaRoleStatus.SchemaStatus).Should(HaveLen(5))
-
-			Expect(schemaUserClaimStatus.SchemaRoleStatus.SchemaStatus["schema1"]).Should(Equal("valid"))
-			Expect(schemaUserClaimStatus.SchemaRoleStatus.SchemaStatus["schema2"]).Should(Equal("valid"))
-			Expect(schemaUserClaimStatus.SchemaRoleStatus.SchemaStatus["schema3"]).Should(Equal("valid"))
-			Expect(schemaUserClaimStatus.SchemaRoleStatus.SchemaStatus["schema4"]).Should(Equal("valid"))
-			Expect(schemaUserClaimStatus.SchemaRoleStatus.SchemaStatus["public"]).Should(Equal("valid"))
-
-			Expect(schemaUserClaimStatus.SchemaRoleStatus.RoleStatus["schema1_regular"]).Should(Equal("valid"))
-			Expect(schemaUserClaimStatus.SchemaRoleStatus.RoleStatus["schema2_admin"]).Should(Equal("valid"))
-			Expect(schemaUserClaimStatus.SchemaRoleStatus.RoleStatus["schema3_readonly"]).Should(Equal("valid"))
-			Expect(schemaUserClaimStatus.SchemaRoleStatus.RoleStatus["schema4_admin"]).Should(Equal("valid"))
-			Expect(schemaUserClaimStatus.SchemaRoleStatus.RoleStatus["public_admin"]).Should(Equal("valid"))
-
-			//-----------------
-			exists, err := dbClient.SchemaExists("schema1")
-			Expect(exists).Should(BeTrue())
-			Expect(err).Should(BeNil())
-
-			exists, err = dbClient.RoleExists("schema1_regular")
-			Expect(exists).Should(BeTrue())
-			Expect(err).Should(BeNil())
-			exists, err = dbClient.UserExists("testclaim_user_a")
-			Expect(exists).Should(BeTrue())
-			Expect(err).Should(BeNil())
-			//-----------------
-			exists, err = dbClient.SchemaExists("schema2")
-			Expect(exists).Should(BeTrue())
-			Expect(err).Should(BeNil())
-
-			exists, err = dbClient.RoleExists("schema2_admin")
-			Expect(exists).Should(BeTrue())
-			Expect(err).Should(BeNil())
-			//-----------------
-			exists, err = dbClient.SchemaExists("schema3")
-			Expect(exists).Should(BeTrue())
-			Expect(err).Should(BeNil())
-
-			exists, err = dbClient.RoleExists("schema3_readonly")
-			Expect(exists).Should(BeTrue())
-			Expect(err).Should(BeNil())
-			//----------PUBLIC-------
-			exists, err = dbClient.RoleExists("public_admin")
-			Expect(exists).Should(BeTrue())
-			Expect(err).Should(BeNil())
-
-		})
-	}
-}
-
-func TestSchemaUserClaimReconcile_WithNewUserSchemasRoles_UpdatePassword(t *testing.T) {
-	RegisterFailHandler(Fail)
-	logf.SetLogger(zap.New(zap.WriteTo(GinkgoWriter), zap.UseDevMode(true)))
-
-	testDB := SetupSqlDB(t, "user_b", "masterpassword")
-	defer testDB.Close()
-
-	type reconciler struct {
-		Client             client.Client
-		Log                logr.Logger
-		Scheme             *runtime.Scheme
-		Config             *roleclaim.RoleConfig
-		DbIdentifierPrefix string
-		Context            context.Context
-		Request            controllerruntime.Request
-	}
-
-	viperObj := viper.New()
-	viperObj.Set("passwordconfig::passwordRotationPeriod", 60)
-
-	tests := []struct {
-		name    string
-		rec     reconciler
-		wantErr bool
-	}{
-		{
-			"Get UserSchema claim 2",
-			reconciler{
-				Client: &MockClient{Port: strconv.Itoa(testDB.Port)},
-				Config: &roleclaim.RoleConfig{
-					Viper: viperObj,
-					Class: "default",
-				},
-				Request: controllerruntime.Request{
-					NamespacedName: types.NamespacedName{Namespace: "schema-user-test", Name: "schema-user-claim-2"},
-				},
-				Log: zap.New(zap.UseDevMode(true)),
-			},
-			false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			r := &DbRoleClaimReconciler{
-				Client: tt.rec.Client,
-				Config: tt.rec.Config,
-			}
-
-			r.reconciler = &roleclaim.DbRoleClaimReconciler{
-				Client: r.Client,
-				Config: r.Config,
-			}
-
-			existingDBConnInfo, err := persistancev1.ParseUri(testDB.URL())
-			if err != nil {
-				t.Errorf("error = %v", err)
-				return
-			}
-
-			dbClient, err := basefun.GetClientForExistingDB(existingDBConnInfo, &controllerruntime.Log)
-			if err != nil {
-				t.Errorf("error = %v", err)
-				return
-			}
-
-			//seed database to simulate existing user
-			dbClient.CreateSchema("schema1")
-			dbClient.CreateRegularRole(existingDBConnInfo.DatabaseName, "schema1_admin", "schema1")
-			dbClient.CreateUser("user2_b", "schema1_admin", "123")
-
-			result, err := r.reconciler.Reconcile(tt.rec.Context, tt.rec.Request)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			Expect(result.Requeue).Should(BeFalse())
-
-			var responseUpdate = r.Client.(*MockClient).GetResponseUpdate()
-			Expect(responseUpdate).Should(Not(BeNil()))
-			var schemaUserClaimStatus = responseUpdate.(*persistancev1.DbRoleClaim).Status
-			Expect(schemaUserClaimStatus).Should(Not(BeNil()))
-			Expect(schemaUserClaimStatus.Error).Should(BeEmpty())
-			Expect(schemaUserClaimStatus.Username).Should(Equal("testclaim_user_b"))
-			Expect(schemaUserClaimStatus.SchemaRoleStatus.SchemaStatus).Should(HaveLen(3))
-
-			Expect(schemaUserClaimStatus.SchemaRoleStatus.SchemaStatus["schema1"]).Should(Equal("valid"))
-			Expect(schemaUserClaimStatus.SchemaRoleStatus.SchemaStatus["schema2"]).Should(Equal("valid"))
-			Expect(schemaUserClaimStatus.SchemaRoleStatus.SchemaStatus["schema3"]).Should(Equal("valid"))
-
-			Expect(schemaUserClaimStatus.SchemaRoleStatus.RoleStatus["schema1_regular"]).Should(Equal("valid"))
-			Expect(schemaUserClaimStatus.SchemaRoleStatus.RoleStatus["schema2_admin"]).Should(Equal("valid"))
-			Expect(schemaUserClaimStatus.SchemaRoleStatus.RoleStatus["schema3_readonly"]).Should(Equal("valid"))
-
-			//-----------------
-			exists, err := dbClient.SchemaExists("schema1")
-			Expect(exists).Should(BeTrue())
-			Expect(err).Should(BeNil())
-
-			exists, err = dbClient.RoleExists("schema1_regular")
-			Expect(exists).Should(BeTrue())
-			Expect(err).Should(BeNil())
-			exists, err = dbClient.UserExists("testclaim_user_b")
-			Expect(exists).Should(BeTrue())
-			Expect(err).Should(BeNil())
-			//-----------------
-			exists, err = dbClient.SchemaExists("schema2")
-			Expect(exists).Should(BeTrue())
-			Expect(err).Should(BeNil())
-
-			exists, err = dbClient.RoleExists("schema2_admin")
-			Expect(exists).Should(BeTrue())
-			Expect(err).Should(BeNil())
-			//-----------------
-			exists, err = dbClient.SchemaExists("schema3")
-			Expect(exists).Should(BeTrue())
-			Expect(err).Should(BeNil())
-
-			exists, err = dbClient.RoleExists("schema3_readonly")
-			Expect(exists).Should(BeTrue())
-			Expect(err).Should(BeNil())
 		})
 	}
 }
